code/serving_dummy/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers>=2.2.0 # Use a reasonably recent version
faiss-cpu
numpy
pydantic<2.0 # Pinning below 2.0 as older ST often needed this
torch
prometheus-fastapi-instrumentator
jinja2
prometheus-client
python-multipart
PyMuPDF
optimum[onnxruntime]
onnx